# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateInterviewRequest": {
        "properties": {
          "candidateName": {
            "anyOf": [
              {
                "maxLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "면접관 화면에 표시할 지원자 이름. 비우면 기본 라벨을 쓴다.",
            "examples": [
              "김지원"
            ],
            "title": "Candidatename"
          }
        },
        "title": "CreateInterviewRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "OpenAPI 문서용. 실제 직렬화는 error_body() 가 한다.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterviewListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InterviewListItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "InterviewListResponse",
        "type": "object"
      },
      "InterviewResponse": {
        "properties": {
          "candidateName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidatename"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "interviewId": {
            "title": "Interviewid",
            "type": "string"
          },
          "interviewerId": {
            "title": "Interviewerid",
            "type": "string"
          }
        },
        "required": [
          "interviewId",
          "interviewerId",
          "candidateName",
          "createdAt"
        ],
        "title": "InterviewResponse",
        "type": "object"
      },
...
      "ReviewUpdateRequest": {
        "description": "`PATCH /interviews/{interviewId}` — 검토 상태 · 메모 (#137 1-3). 둘 다 선택이다.\n\nnull 은 보내지 않은 것과 같다. 메모를 비우려면 빈 문자열을 보낸다. 길이와 NUL 은\n경계에서 막는다 — PostgreSQL TEXT 는 NUL 을 못 넣어 500 이 된다.",
        "properties": {
          "memo": {
            "anyOf": [
              {
                "maxLength": 4000,
                "pattern": "^[^\\x00]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memo"
          },
          "reviewStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ReviewUpdateRequest",
        "type": "object"
      },
      "ReviewUpdateResponse": {
        "properties": {
          "memo": {
            "title": "Memo",
            "type": "string"
          },
          "reviewStatus": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "reviewedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "CONFIRMED 가 된 시각. 다른 상태로 돌아가면 null.",
            "title": "Reviewedat"
          }
        },
        "required": [
          "reviewStatus",
          "reviewedAt",
          "memo"
        ],
        "title": "ReviewUpdateResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "IRYA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/interviews": {
      "get": {
        "description": "내가 만든 면접을 최신순으로 준다. 필터 · 정렬은 FE 가 한다 (#137 1-1).\n\n면접관은 늘 나라서 `interviewer` 는 토큰의 사용자고, 직무는 내 컨텍스트에서 온다.",
        "operationId": "list_interviews_api_v1_interviews_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "내 면접 목록",
        "tags": [
          "면접"
        ]
      },
      "post": {
        "description": "면접관이 새로운 면접 정보를 생성한다.\n\nSession 과 LiveKit Room 은 여기서 만들지 않는다 —\n`POST /interviews/{interviewId}/sessions` 가 담당한다.\n\n면접의 주인(`interviewerId`)은 토큰의 사용자다. 요청 본문으로 받으면 아무 이름으로나\n면접을 만들 수 있어, 그 면접을 주인만 다룬다는 판단이 서지 않는다.",
        "operationId": "create_interview_api_v1_interviews_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "요청값 검증 실패"
          }
        },
        "summary": "면접 생성",
        "tags": [
          "면접"
        ]
      }
    },
    "/api/v1/interviews/{interviewId}": {
      "get": {
        "description": "생성된 면접의 기본 정보를 조회한다.",
        "operationId": "get_interview_api_v1_interviews__interviewId__get",
        "parameters": [
          {
            "in": "path",
            "name": "interviewId",
            "required": true,
            "schema": {
              "title": "Interviewid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "면접을 찾을 수 없음 (남의 면접 포함)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "면접 조회",
        "tags": [
          "면접"
        ]
      },
      "patch": {
        "description": "검토 상태와 메모를 바꾼다. 보낸 것만 바뀐다 (#137 1-3).\n\n`CONFIRMED` 가 될 때 `reviewedAt` 을 찍고 다른 상태로 돌아가면 비운다. AI 결과가\n없어도 된다 — 검토는 면접 단위다.",
        "operationId": "update_review_api_v1_interviews__interviewId__patch",
        "parameters": [
          {
            "in": "path",
            "name": "interviewId",
            "required": true,
            "schema": {
              "title": "Interviewid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "면접을 찾을 수 없음 (남의 면접 포함)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "검토 상태 · 메모 수정",
        "tags": [
          "면접"
        ]
      }
    },
    "/api/v1/interviews/{interviewId}/resume": {
      "post": {
        "description": "지원자 이력서를 올린다. **PDF 만 받는다.**\n\n**면접 한 건에 한 장이고 다시 올리면 덮어쓴다.** FE 가 목록도 삭제도 두지 않은\n것이 그 전제다(#81) — 새 이력서를 올리면 앞의 것은 쓸 일이 없다.\n\n응답은 기업 컨텍스트 문서와 같은 모양(`ContextDoc`)이다. FE 가 같은 카드\n컴포넌트로 그린다.\n\n본문은 기업 컨텍스트 문서와 같이 응답 뒤에 따로 뽑는다(#143). 응답은 `parsing`\n이다. 꼬리질문과 리포트가 이 본문을 쓴다(#70).",
        "operationId": "upload_resume_api_v1_interviews__interviewId__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "interviewId",
            "required": true,
            "schema": {
              "title": "Interviewid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_api_v1_interviews__interviewId__resume_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextDocResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "면접을 찾을 수 없음 (남의 면접 포함)"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "파일이 너무 큼"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "지원하지 않는 형식"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "지원자 이력서 업로드",
        "tags": [
          "면접"
        ]
      }
    },
    "/api/v1/interviews/{interviewId}/review": {
      "get": {
        "description": "면접이 끝난 뒤의 검토 화면 — 요약 · coverage · 타임라인 · 검토 항목 (#137 1-2).\n\n기준 세션은 그 면접에서 가장 나중에 끝난 세션이다(목록과 같다). 끝난 세션이\n없거나 요약이 아직 PROCESSING 이면 202 다. FE 는 그동안 다시 조회한다.\n\n**요약이 FAILED 여도 200 이다** (#163). `summary` 는 null 이고 moments ·\nfindings 는 비며 coverage 는 역량 전부 MISSING 이다. 메모 · 검토 확정은 AI\n결과와 상관없이 할 수 있어야 한다.",
        "operationId": "get_review_api_v1_interviews__interviewId__review_get",
        "parameters": [
          {
            "in": "path",
            "name": "interviewId",
            "required": true,
            "schema": {
              "title": "Interviewid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewProcessingResponse"
                }
              }
            },
            "description": "준비 중"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "면접을 찾을 수 없음 (남의 면접 포함)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "면접 기록 조회",
        "tags": [
          "면접"
        ]
      }
    },
    "/api/v1/interviews/{interviewId}/review/marks/{itemId}": {
      "put": {
        "description": "검토 항목(finding)의 채택 · 반려, 문답(moment)의 북마크 (#137 1-4).\n\n`itemId` 는 지금 상세의 `findings[].id` 나 `moments[].id` 여야 한다. 아니면\n404 다 — 아무 id 로 행이 쌓이지 않고, 분석이 READY 가 되기 전에는 표시할 것도\n없다 (#163 결정). 문답은 북마크만, 검토 항목은 채택만 받는다(1-4 「둘 중\n하나」). 표시는 기준 세션에 남는다.",
        "operationId": "put_mark_api_v1_interviews__interviewId__review_marks__itemId__put",
        "parameters": [
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "title": "Itemid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "interviewId",
            "required": true,
            "schema": {
              "title": "Interviewid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "면접이 없거나 지금 상세에 그 항목이 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "검토 항목 채택 · 문답 북마크",
        "tags": [
          "면접"
        ]
      }
    },
    "/api/v1/interviews/{interviewId}/sessions": {
      "post": {
        "description": "생성된 면접에 실제 화상면접 Session 을 만들고 지원자 초대 링크를 발급한다.\n\nLiveKit Room 을 미리 만든다. 입장 시 자동 생성되기는 하지만\n`max_participants` 같은 설정을 적용하려면 사전 생성이 필요하다.",
        "operationId": "create_session_api_v1_interviews__interviewId__sessions_post",
        "parameters": [
          {
            "in": "path",
            "name": "interviewId",
            "required": true,
            "schema": {
              "title": "Interviewid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "면접을 찾을 수 없음 (남의 면접 포함)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "면접 Session 생성",
        "tags": [
          "면접"
        ]
      }
    },
...
  }
}